A bit of refactoring
Since we're calling Teng methods directly, change this by creating new
methods in BookList::DB that will serve as our API for the database.
This should allow, for example. changing from Teng to some other DBI
wrapper or ORM, and further separate concerns between the Books
controller and the database.